+++ src/compliment/sources/local_bindings.clj
@@ -1,30 +1,65 @@
 (ns compliment.sources.local-bindings
   "Completion source for local bindings introduced by defn, let and the like."
   (:require [compliment.sources :refer [defsource]]
-            [compliment.sources.ns-mappings :refer [var-symbol? dash-matches?]]))
+            [compliment.sources.ns-mappings :refer [var-symbol? dash-matches?]]
+            [compliment.utils :as utils]))
 
-(def let-like-forms '#{let if-let when-let if-some when-some loop with-open
-                       dotimes with-local-vars})
+(defn- let-like-form? [x]
+  (and (symbol? x)
+       (contains? #{"let"
+                    "if-let"
+                    "when-let"
+                    "if-some"
+                    "when-some"
+                    "loop"
+                    "with-open"
+                    "dotimes"
+                    "with-local-vars"}
+                  (name x))))
 
-(def defn-like-forms '#{defn defn- fn defmacro defmethod})
+(defn- defn-like-form? [x]
+  (and (symbol? x)
+       (contains? #{"defn"
+                    "defn-"
+                    "fn"
+                    "fn*"
+                    "defmacro"
+                    "defmethod"}
+                  (name x))))
 
-(def doseq-like-forms '#{doseq for})
+(defn- doseq-like-form? [x]
+  (and (symbol? x)
+       (contains? #{"doseq" "for"} (name x))))
 
-(def letfn-like-forms '#{letfn})
+(defn- letfn-like-form? [x]
+  (and (symbol? x)
+       (contains? #{"letfn"} (name x))))
 
 (def destructuring-key-names #{"keys" "strs" "syms"})
+
+(defn extract-tag-from-bound-to [ns bound-to]
+  (when bound-to
+    (when-let [found (or (-> bound-to meta :tag)
+                         (utils/var->class ns bound-to)
+                         (utils/invocation-form->class ns bound-to))]
+      (if (class? found)
+        (-> ^Class found .getName symbol)
+        found))))
 
 (defn parse-binding
   "Given a binding node returns the list of local bindings introduced by that
   node. Handles vector and map destructuring."
-  [binding-node]
+  [ns binding-node bound-to]
   (cond (vector? binding-node)
-        (mapcat parse-binding binding-node)
+        (mapcat (fn [x]
+                  (parse-binding ns x nil))
+                binding-node)
 
         (map? binding-node)
         (let [normal-binds (->> (keys binding-node)
                                 (remove keyword?)
-                                (mapcat parse-binding))
+                                (mapcat (fn [x]
+                                          (parse-binding ns x nil))))
               keys-binds (->> binding-node
                               (mapcat (fn [[k v]]
                                         ;; This handles both plain and
@@ -38,17 +73,22 @@
             as-bind (conj as-bind)))
 
         (not (#{'& '_} binding-node))
-        [binding-node]))
+        [(if (-> binding-node meta :tag)
+           binding-node
+           (if-let [candidate (extract-tag-from-bound-to ns bound-to)]
+             (vary-meta binding-node assoc :tag candidate)
+             binding-node))]))
 
 (defn parse-fn-body
   "Extract function name and arglists from the function body, return list of all
   completable variables."
-  [fn-body]
+  [ns fn-body]
   (let [fn-name (when (symbol? (first fn-body))
                   (first fn-body))
         fn-body (if fn-name (rest fn-body) fn-body)]
     (cond->
-        (mapcat parse-binding
+        (mapcat (fn [x]
+                  (parse-binding ns x nil))
                 (loop [[c & r] fn-body, bnodes []]
                   (cond (nil? c) bnodes
                         (list? c) (recur r (conj bnodes (first c))) ;; multi-arity case
@@ -60,25 +100,30 @@
   "When given a form that has a binding vector traverses that binding vector and
   returns the list of all local bindings."
   [form ns]
-  (when (list? form)
+  (when (seq? form)
     (let [sym (first form)
           locals-meta (when (symbol? sym)
                         (:completion/locals (meta (ns-resolve ns sym))))]
-      (cond (or (let-like-forms sym) (= locals-meta :let))
-            (mapcat parse-binding (take-nth 2 (second form)))
+      (cond (or (let-like-form? sym) (= locals-meta :let))
+            (mapcat (fn [[x y]]
+                      (parse-binding ns x y))
+                    (partition 2 (second form)))
 
-            (or (defn-like-forms sym) (= locals-meta :defn))
-            (parse-fn-body (rest form))
+            (or (defn-like-form? sym) (= locals-meta :defn))
+            (parse-fn-body ns (rest form))
 
-            (or (letfn-like-forms sym) (= locals-meta :letfn))
-            (mapcat parse-fn-body (second form))
+            (or (letfn-like-form? sym) (= locals-meta :letfn))
+            (mapcat (fn [fn-body]
+                      (parse-fn-body ns fn-body))
+                    (second form))
 
-            (or (doseq-like-forms sym) (= locals-meta :doseq))
+            (or (doseq-like-form? sym) (= locals-meta :doseq))
             (->> (partition 2 (second form))
                  (mapcat (fn [[left right]]
                            (if (= left :let)
                              (take-nth 2 right) [left])))
-                 (mapcat parse-binding))
+                 (mapcat (fn [x]
+                           (parse-binding ns x nil))))
 
             (= sym 'as->) [(nth form 2)]))))